Remove dependency of init on allvars and move allvars to run.c
