+++ libgadget/runtests.c
@@ -89,7 +89,17 @@ void runtests(int RestartSnapNum)
     DomainDecomp ddecomp[1] = {0};
     /* So we can run a test on the final snapshot*/
     All.TimeMax = All.TimeInit * 1.1;
-    inttime_t Ti_Current = init(RestartSnapNum, ddecomp);          /* ... read in initial model */
+    double MeanSeparation[6] = {0};
+    get_mean_separation(MeanSeparation, All.BoxSize, All.NTotalInit);
+
+    inttime_t Ti_Current = init(RestartSnapNum, All.TimeIC, All.TimeInit, All.TimeMax, &All.CP, All.SnapshotWithFOF, All.MassiveNuLinRespOn, All.BoxSize, All.G, All.MassTable, MeanSeparation);
+
+    domain_decompose_full(ddecomp);	/* do initial domain decomposition (gives equal numbers of particles) */
+
+    if(All.DensityOn) {
+        double uu_in_cgs = All.UnitEnergy_in_cgs / All.UnitMass_in_g;
+        setup_smoothinglengths(RestartSnapNum, ddecomp, &All.CP, All.BlackHoleOn, All.BoxSize, All.MinEgySpec, uu_in_cgs, Ti_Current, All.TimeInit, MeanSeparation[0]);
+    }
 
     struct IOTable IOTable = {0};
     register_io_blocks(&IOTable, 0);
@@ -194,7 +204,19 @@ runfof(int RestartSnapNum)
     gravpm_init_periodic(&pm, All.BoxSize, All.Asmth, All.Nmesh, All.G);
     DomainDecomp ddecomp[1] = {0};
     /* ... read in initial model */
-    DriftKickTimes times = init_driftkicktime(init(RestartSnapNum, ddecomp));
+    double MeanSeparation[6] = {0};
+    get_mean_separation(MeanSeparation, All.BoxSize, All.NTotalInit);
+
+    inttime_t Ti_Current = init(RestartSnapNum, All.TimeIC, All.TimeInit, All.TimeMax, &All.CP, All.SnapshotWithFOF, All.MassiveNuLinRespOn, All.BoxSize, All.G, All.MassTable, MeanSeparation);
+
+    domain_decompose_full(ddecomp);	/* do initial domain decomposition (gives equal numbers of particles) */
+
+    if(All.DensityOn) {
+        double uu_in_cgs = All.UnitEnergy_in_cgs / All.UnitMass_in_g;
+        setup_smoothinglengths(RestartSnapNum, ddecomp, &All.CP, All.BlackHoleOn, All.BoxSize, All.MinEgySpec, uu_in_cgs, Ti_Current, All.TimeInit, MeanSeparation[0]);
+    }
+
+    DriftKickTimes times = init_driftkicktime(Ti_Current);
     /*FoF needs a tree*/
     int HybridNuGrav = All.HybridNeutrinosOn && All.Time <= All.HybridNuPartTime;
     /* Regenerate the star formation rate for the FOF table.*/
@@ -229,7 +251,18 @@ runpower(int RestartSnapNum)
     PetaPM pm = {0};
     gravpm_init_periodic(&pm, All.BoxSize, All.Asmth, All.Nmesh, All.G);
     DomainDecomp ddecomp[1] = {0};
-    init(RestartSnapNum, ddecomp);          /* ... read in initial model */
+    /* ... read in initial model */
+    double MeanSeparation[6] = {0};
+    get_mean_separation(MeanSeparation, All.BoxSize, All.NTotalInit);
+
+    inttime_t Ti_Current = init(RestartSnapNum, All.TimeIC, All.TimeInit, All.TimeMax, &All.CP, All.SnapshotWithFOF, All.MassiveNuLinRespOn, All.BoxSize, All.G, All.MassTable, MeanSeparation);
+
+    domain_decompose_full(ddecomp);	/* do initial domain decomposition (gives equal numbers of particles) */
+
+    if(All.DensityOn) {
+        double uu_in_cgs = All.UnitEnergy_in_cgs / All.UnitMass_in_g;
+        setup_smoothinglengths(RestartSnapNum, ddecomp, &All.CP, All.BlackHoleOn, All.BoxSize, All.MinEgySpec, uu_in_cgs, Ti_Current, All.TimeInit, MeanSeparation[0]);
+    }
 
     /*PM needs a tree*/
     ForceTree Tree = {0};
